% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_FRD_data}
\alias{get_FRD_data}
\title{Download FRD data}
\usage{
get_FRD_data(granularity = 1)
}
\arguments{
\item{granularity}{Numeric. The desired granularity for the ouput, 1 for new regions,
2 for older regions and 3 for department.}
}
\value{
A tibble containing FRD data for the desired granularity
}
\description{
Download FRD data
}
