% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_HMD_data}
\alias{get_HMD_data}
\title{Download HMD data}
\usage{
get_HMD_data(countries, username, password)
}
\arguments{
\item{countries}{A vector of country names. Names, if provided will be used
as Country IDs in the returned tibble.}

\item{username}{a valid HMD user name}

\item{password}{a valid HMD password}
}
\value{
A tibble containing HMD data for the desired countries
}
\description{
Download HMD data
}
