% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_STMF_data}
\alias{get_STMF_data}
\title{Download STMF data}
\usage{
get_STMF_data(username, password)
}
\arguments{
\item{username}{a valid HMD user name}

\item{password}{a valid HMD password}
}
\value{
A tibble containing HMD Short Term Mortality Fluctuations (STMF) data
}
\description{
Download STMF data
}
