% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impmort-package.R
\docType{package}
\name{impmort-package}
\alias{impmort-package}
\title{impmort: Import Free Access Data For Mortality Modelling.}
\description{
Import mortality data from more than 41 countries using HMD data or the
French regional mortality (FRD) data with 3 levels of granularity. Interface
is very intuitive you just have to register on HMD and get your username and
password.
}
